backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-appconf==1.0.5
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
node==1.2.1
odict==1.9.0
pilkit==2.0
Pillow==9.5.0
plumber==1.7
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
